+++ NEWS.md
@@ -28,6 +28,7 @@ based on the discussions in the meeting early February.
 * Both XLSX sheets 'META-Room' and 'META-Variable' have a column for optional information
     about the measurement location (e.g., next to window, center of room, and any other
     important/helpful information) or variable (optional meta data).
+* Additional tinytests for support functions.
 
 
 # annex 0.2-0